refactor(Shared.Core): nest ToolError under Shared.Core; create namespace anchor; rename file
### New Shared.Core sub-namespace

Adds the missing Shared.Core sub-namespace anchor (analogous to Sample.Core
declared in SharedConstants/Sample.swift). Anchor file Shared.Core.swift.

### Type relocation

- ToolError (file-scope public enum) -> Shared.Core.ToolError

### File rename

- Sources/Shared/Core/ToolError.swift -> Sources/Shared/Core/Shared.Core.ToolError.swift
- New: Sources/Shared/Core/Shared.Core.swift (namespace anchor)

### Target dependency

SharedCore target gains 'SharedConstants' dep — needed to see the Shared
namespace declared in SharedConstants/Shared.swift. Shared was already in
practice a transitive dep through everything; making it explicit keeps the
graph honest.

### External caller sweep (7 files)

- Sources/SearchToolProvider/CompositeToolProvider.swift
- Sources/Services/Services.ServiceContainer.swift
- Sources/MCP/SharedTools/MCP.SharedTools.ArgumentExtractor.swift
- Sources/MCP/Support/MCP.Support.DocsResourceProvider.swift
- Tests/CLICommandTests/ServeTests/ServeTests.swift
- Tests/SearchToolProviderTests/CupertinoSearchToolProviderTests.swift
- Tests/MCP/SharedToolsTests/ArgumentExtractorTests.swift

All ToolError references qualified through Shared.Core.

### Verification

- xcrun swift build (clean): 0 warnings, 0 errors
- xcrun swift test (clean):  1300/1300 pass
